+++ NEWS.md
@@ -1,5 +1,6 @@
 # spant 3.4.0
 * Added a function to read DKD's sLASER MoCo log file.
+* Bug fix for TE parameter determination from Siemens IMA files.
 
 # spant 3.3.0
 * Added voxel render on MRI images to fit_svs.